+++ src/AdminPaymentFormPostType.php
@@ -3,14 +3,22 @@
 /**
  * Title: WordPress admin payment form post type
  * Description:
- * Copyright: Copyright (c) 2005 - 2016
+ * Copyright: Copyright (c) 2005 - 2017
  * Company: Pronamic
  *
  * @author Remco Tolsma
- * @version 1.4.7
+ * @version 1.6.0
  * @since 1.0.0
  */
 class Pronamic_WP_Pay_Extensions_GravityForms_AdminPaymentFormPostType {
+	/**
+	 * Post type
+	 */
+	const POST_TYPE = 'pronamic_pay_gf';
+
+	/**
+	 * Constructs and intialize admin payment form post type.
+	 */
 	public function __construct() {
 		add_filter( 'manage_edit-pronamic_pay_gf_columns', array( $this, 'edit_columns' ) );
 
@@ -18,7 +26,13 @@ class Pronamic_WP_Pay_Extensions_GravityForms_AdminPaymentFormPostType {
 
 		add_action( 'add_meta_boxes', array( $this, 'add_meta_boxes' ) );
 
-		add_action( 'save_post', array( $this, 'save_post' ) );
+		if ( version_compare( GFCommon::$version, '1.7', '>=' ) ) {
+			add_action( 'gform_after_delete_form', array( $this, 'delete_payment_form' ) );
+		}
+
+		add_filter( 'wp_insert_post_data', array( $this, 'insert_post_data' ), 99, 2 );
+
+		add_action( 'save_post_' . self::POST_TYPE, array( $this, 'save_post' ) );
 	}
 
 	public function edit_columns( $columns ) {
@@ -90,7 +104,67 @@ class Pronamic_WP_Pay_Extensions_GravityForms_AdminPaymentFormPostType {
 	 * @param WP_Post $post The object for the current post/page.
 	 */
 	public function meta_box_config( $post ) {
-		include dirname( __FILE__ ) . '/../views/html-admin-meta-box-config.php';
+		$form_id = get_post_meta( $post->ID, '_pronamic_pay_gf_form_id', true );
+		$post_id = $post->ID;
+
+		include dirname( __FILE__ ) . '/../views/html-admin-feed-meta-box.php';
+	}
+
+	/**
+	 * When the form is deleted from the trash, deletes our custom post.
+	 *
+	 * @param int $form_id The ID of the form being deleted.
+	 */
+	public function delete_payment_form( $form_id ) {
+		$query = new WP_Query( array(
+			'post_type'			=> 'pronamic_pay_gf',
+			'meta_key'			=> array(
+				'key'			=> '_pronamic_pay_gf_form_id',
+				'value'			=> $form_id,
+			),
+		) );
+
+		foreach ( $query->posts as $post ) {
+			wp_delete_post( $post->ID, true );
+		}
+	}
+
+	public function insert_post_data( $data, $postarr ) {
+		// Check if pay feed post type
+		if ( 'pronamic_pay_gf' !== $postarr['post_type'] ) {
+			return $data;
+		}
+
+		// Check if our nonce is set.
+		if ( ! filter_has_var( INPUT_POST, 'pronamic_pay_nonce' ) ) {
+			return $data;
+		}
+
+		$nonce = filter_input( INPUT_POST, 'pronamic_pay_nonce', FILTER_SANITIZE_STRING );
+
+		// Verify that the nonce is valid.
+		if ( ! wp_verify_nonce( $nonce, 'pronamic_pay_save_pay_gf' ) ) {
+			return $data;
+		}
+
+		// If this is an autosave, our form has not been submitted, so we don't want to do anything.
+		if ( defined( 'DOING_AUTOSAVE' ) && DOING_AUTOSAVE ) {
+			return $data;
+		}
+
+		// Check the user's permissions.
+		if ( ! current_user_can( 'edit_post', $postarr['ID'] ) ) {
+			return $data;
+		}
+
+		/* OK, its safe for us to save the data now. */
+		if ( filter_has_var( INPUT_POST, '_pronamic_pay_gf_post_title' ) ) {
+			$post_title = filter_input( INPUT_POST, '_pronamic_pay_gf_post_title', FILTER_SANITIZE_STRING );
+
+			$data['post_title'] = sanitize_text_field( wp_unslash( $post_title ) );
+		}
+
+		return $data;
 	}
 
 	/**
@@ -101,58 +175,61 @@ class Pronamic_WP_Pay_Extensions_GravityForms_AdminPaymentFormPostType {
 	public function save_post( $post_id ) {
 		// Check if our nonce is set.
 		if ( ! filter_has_var( INPUT_POST, 'pronamic_pay_nonce' ) ) {
-			return $post_id;
+			return;
 		}
 
 		$nonce = filter_input( INPUT_POST, 'pronamic_pay_nonce', FILTER_SANITIZE_STRING );
 
 		// Verify that the nonce is valid.
 		if ( ! wp_verify_nonce( $nonce, 'pronamic_pay_save_pay_gf' ) ) {
-			return $post_id;
+			return;
 		}
 
 		// If this is an autosave, our form has not been submitted, so we don't want to do anything.
 		if ( defined( 'DOING_AUTOSAVE' ) && DOING_AUTOSAVE ) {
-			return $post_id;
+			return;
 		}
 
 		// Check the user's permissions.
-		if ( 'page' === get_post_type( $post_id ) ) {
-			if ( ! current_user_can( 'edit_page', $post_id ) ) {
-				return $post_id;
-			}
-		} else {
-			if ( ! current_user_can( 'edit_post', $post_id ) ) {
-				return $post_id;
-			}
+		if ( ! current_user_can( 'edit_post', $post_id ) ) {
+			return;
 		}
 
 		/* OK, its safe for us to save the data now. */
 		$definition = array(
-			'_pronamic_pay_gf_form_id'                            => 'sanitize_text_field',
-			'_pronamic_pay_gf_config_id'                          => 'sanitize_text_field',
-			'_pronamic_pay_gf_entry_id_prefix'                    => 'sanitize_text_field',
-			'_pronamic_pay_gf_transaction_description'            => 'sanitize_text_field',
-			'_pronamic_pay_gf_condition_enabled'                  => FILTER_VALIDATE_BOOLEAN,
-			'_pronamic_pay_gf_condition_field_id'                 => 'sanitize_text_field',
-			'_pronamic_pay_gf_condition_operator'                 => 'sanitize_text_field',
-			'_pronamic_pay_gf_condition_value'                    => 'sanitize_text_field',
-			'_pronamic_pay_gf_delay_admin_notification'           => FILTER_VALIDATE_BOOLEAN,
-			'_pronamic_pay_gf_delay_user_notification'            => FILTER_VALIDATE_BOOLEAN,
-			'_pronamic_pay_gf_delay_notification_ids'             => array(
+			'_pronamic_pay_gf_form_id'                      => 'sanitize_text_field',
+			'_pronamic_pay_gf_config_id'                    => 'sanitize_text_field',
+			'_pronamic_pay_gf_entry_id_prefix'              => 'sanitize_text_field',
+			'_pronamic_pay_gf_transaction_description'      => 'sanitize_text_field',
+			'_pronamic_pay_gf_condition_enabled'            => FILTER_VALIDATE_BOOLEAN,
+			'_pronamic_pay_gf_condition_field_id'           => 'sanitize_text_field',
+			'_pronamic_pay_gf_condition_operator'           => 'sanitize_text_field',
+			'_pronamic_pay_gf_condition_value'              => 'sanitize_text_field',
+			'_pronamic_pay_gf_delay_admin_notification'     => FILTER_VALIDATE_BOOLEAN,
+			'_pronamic_pay_gf_delay_user_notification'      => FILTER_VALIDATE_BOOLEAN,
+			'_pronamic_pay_gf_delay_notification_ids'       => array(
 				'filter'    => FILTER_SANITIZE_STRING,
 				'flags'     => FILTER_REQUIRE_ARRAY,
 			),
-			'_pronamic_pay_gf_delay_post_creation'                => FILTER_VALIDATE_BOOLEAN,
-			'_pronamic_pay_gf_fields'                             => array(
+			'_pronamic_pay_gf_delay_post_creation'          => FILTER_VALIDATE_BOOLEAN,
+			'_pronamic_pay_gf_fields'                       => array(
 				'filter'    => FILTER_SANITIZE_STRING,
 				'flags'     => FILTER_REQUIRE_ARRAY,
 			),
 			'_pronamic_pay_gf_links' => array(
 				'filter'    => FILTER_SANITIZE_STRING,
 				'flags'     => FILTER_REQUIRE_ARRAY,
-				),
-			'_pronamic_pay_gf_user_role_field_id'                 => 'sanitize_text_field',
+			),
+			'_pronamic_pay_gf_user_role_field_id'           => 'sanitize_text_field',
+			'_pronamic_pay_gf_subscription_amount_type'     => 'sanitize_text_field',
+			'_pronamic_pay_gf_subscription_amount_field'    => 'sanitize_text_field',
+			'_pronamic_pay_gf_subscription_interval_type'   => 'sanitize_text_field',
+			'_pronamic_pay_gf_subscription_interval'        => FILTER_SANITIZE_NUMBER_INT,
+			'_pronamic_pay_gf_subscription_interval_period' => 'sanitize_text_field',
+			'_pronamic_pay_gf_subscription_interval_field'  => 'sanitize_text_field',
+			'_pronamic_pay_gf_subscription_frequency_type'  => 'sanitize_text_field',
+			'_pronamic_pay_gf_subscription_frequency'       => FILTER_SANITIZE_NUMBER_INT,
+			'_pronamic_pay_gf_subscription_frequency_field' => 'sanitize_text_field',
 		);
 
 		if ( class_exists( 'GFAWeber' ) ) {
@@ -228,6 +305,14 @@ class Pronamic_WP_Pay_Extensions_GravityForms_AdminPaymentFormPostType {
 				update_post_meta( $post_id, $meta_key, $meta_value );
 			} else {
 				delete_post_meta( $post_id, $meta_key );
+			}
+		}
+
+		if ( filter_has_var( INPUT_POST, '_pronamic_pay_gf_condition_field_id' ) ) {
+			if ( '' !== filter_input( INPUT_POST, '_pronamic_pay_gf_condition_field_id' ) ) {
+				update_post_meta( $post_id, '_pronamic_pay_gf_condition_enabled', true );
+			} else {
+				delete_post_meta( $post_id, '_pronamic_pay_gf_condition_enabled' );
 			}
 		}
 	}