+++ src/Entry.php
@@ -3,7 +3,7 @@
 /**
  * Title: WordPress pay extension Gravity Forms entry
  * Description:
- * Copyright: Copyright (c) 2005 - 2016
+ * Copyright: Copyright (c) 2005 - 2017
  * Company: Pronamic
  *
  * @author Remco Tolsma