+++ src/PaymentMethodsField.php
@@ -3,11 +3,11 @@
 /**
  * Title: WordPress pay extension Gravity Forms payment methods
  * Description:
- * Copyright: Copyright (c) 2005 - 2016
+ * Copyright: Copyright (c) 2005 - 2017
  * Company: Pronamic
  *
  * @author Remco Tolsma
- * @version 1.5.2
+ * @version 1.6.0
  * @since 1.4.7
  */
 class Pronamic_WP_Pay_Extensions_GravityForms_PaymentMethodsField extends GF_Field_Select {
@@ -130,8 +130,10 @@ class Pronamic_WP_Pay_Extensions_GravityForms_PaymentMethodsField extends GF_Fie
 			// @todo What todo if error?
 			if ( $field && ! is_wp_error( $this->error ) ) {
 				foreach ( $field['choices'] as $group ) {
-					foreach ( $group['options'] as $value => $label ) {
-						$payment_methods[ $value ] = $label;
+					if ( isset( $group['options'] ) ) {
+						foreach ( $group['options'] as $value => $label ) {
+							$payment_methods[ $value ] = $label;
+						}
 					}
 				}
 			}